--- src/main/java/com/heartz/byeboo/adapter/in/web/dto/QuestDetailResponseDto.java
@@ -1,14 +1,0 @@
-package com.heartz.byeboo.adapter.in.web.dto;
-
-import java.time.LocalDateTime;
-
-public record QuestDetailResponseDto(
-        Long stepNumber,
-        Long questNumber,
-        LocalDateTime createdAt,
-        String question,
-        String answer,
-        String questEmotionState,
-        String imageUrl
-) {
-}
